==> 1_species.Rmd <==
---
title: "species"
author: "jiaqi"
date: "2025-02-13"
output: html_document
---

```{r}
# 加载必要的库
library(dplyr)
library(readr)
```
```{r}
# 读取数据
abundance_data <- read_csv("D:/msc_projecct/OneDrive_2025-02-13/Genome Scan_Alessia 2023/merged_metaphlan_bugs_abundance_table_20211201_version_4_cm.csv")

taxa_annotation <- read_csv("D:/msc_projecct/OneDrive_2025-02-13/Genome Scan_Alessia 2023/taxa_annotation_cm.csv")
```

```{r}
# 筛选 species 或以下分类（含 "s__"）
species_taxa <- taxa_annotation %>%
  filter(grepl("s__", clade_name) & !grepl("t__", clade_name))

head(species_taxa)
View(species_taxa)
```

```{r}
# 提取所有列名
all_columns <- colnames(abundance_data)

head(abundance_data)
```

```{r}
# 找到所有需要保留的列：
# 1. metadata 列（不在 taxa_annotation 里的列）
# 2. species 对应的 taxa 列
metadata_columns <- setdiff(all_columns, taxa_annotation$taxa)  # 非菌群列
species_columns <- species_taxa$taxa  # 仅包含 species 级别的菌群列

head(metadata_columns)
head(species_columns)
length(metadata_columns)
length(species_columns)
```

```{r}
# 组合需要保留的列
final_columns <- c(metadata_columns, species_columns)
head(final_columns)
```

```{r}
# 筛选数据
filtered_data <- abundance_data %>%
  dplyr::select(all_of(final_columns))
```

```{r}
# 查看前几行数据
head(filtered_data)
```
```{r}
# 保存筛选后的数据
write_csv(filtered_data, "D:/msc_projecct/OneDrive_2025-02-13/Genome Scan_Alessia 2023/filtered_abundance_species_nostrain.csv")
```
